Changed the type from normal int to uint16_t

+++ src/fakeMoistSensor.cpp
@@ -6,7 +6,7 @@ FakeMoistSensor::FakeMoistSensor()
     srand(time(NULL));
 }
 
-int FakeMoistSensor::moisture() {
+uint16_t FakeMoistSensor::moisture() {
     // get random value
     return rand() % 100 + 1;
 }
